Fix html anchor parsing wrong in some docs
Some path parse from objects.inv like the one in alembic's would be
'index.html#document-api/ddl#alembic.ddl.mysql.MySQLChangeColumn'.
As you see, there is more than one '#' in this path, and it cannot
be recognised correctly.
That may be the inv file's error. This commit replace the path with
head and tail of the original. It can fix this issue.